Replacing the internal representation of mass/tonnage with doubles

+++ megamek/src/megamek/common/Entity.java
@@ -189,7 +189,7 @@ public abstract class Entity extends TurnOrdered implements Transporter,
      */
     protected String externalId = "-1";
 
-    protected float weight;
+    protected double weight;
     protected boolean omni = false;
     protected String chassis;
     protected String model;
@@ -1087,7 +1087,8 @@ public abstract class Entity extends TurnOrdered implements Transporter,
         this.year = year;
     }
 
-    public float getWeight() {
+    /** @return the tonnage of the Entity, not its weight */
+    public double getWeight() {
         return weight;
     }
 
@@ -1100,7 +1101,7 @@ public abstract class Entity extends TurnOrdered implements Transporter,
         return EntityWeightClass.getClassName(getWeightClass(), this);
     }
 
-    public void setWeight(float weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
         // Any time the weight is reset we need to reset the crew size
         crew.setSize(Compute.getFullCrewSize(this));
@@ -8804,8 +8805,8 @@ public abstract class Entity extends TurnOrdered implements Transporter,
         return false;
     }
 
-    public float getTroopCarryingSpace() {
-        float space = 0;
+    public double getTroopCarryingSpace() {
+        double space = 0;
         for (Transporter t : transports) {
             if (t instanceof TroopSpace) {
                 space += ((TroopSpace) t).totalSpace;
